+++ storage-api/serde-api/src/gen/thrift/gen-php/org/apache/hadoop/hive/serde/Types.php
@@ -142,4 +142,20 @@ $GLOBALS['serde_CONSTANTS']['IntegralTypes'] = array(
   "bigint" => true,
 );
 
+$GLOBALS['serde_CONSTANTS']['MAX_CHAR_LENGTH'] = 255;
+
+$GLOBALS['serde_CONSTANTS']['MAX_VARCHAR_LENGTH'] = 65535;
+
+$GLOBALS['serde_CONSTANTS']['SYSTEM_DEFAULT_PRECISION'] = 38;
+
+$GLOBALS['serde_CONSTANTS']['SYSTEM_DEFAULT_SCALE'] = 18;
+
+$GLOBALS['serde_CONSTANTS']['USER_DEFAULT_PRECISION'] = 10;
+
+$GLOBALS['serde_CONSTANTS']['USER_DEFAULT_SCALE'] = 0;
+
+$GLOBALS['serde_CONSTANTS']['MAX_PRECISION'] = 38;
+
+$GLOBALS['serde_CONSTANTS']['MAX_SCALE'] = 38;
+
