+++ storage-api/serde-api/src/gen/thrift/gen-py/org_apache_hadoop_hive_serde/constants.py
@@ -86,3 +86,11 @@ IntegralTypes = set([
   "int",
   "bigint",
 ])
+MAX_CHAR_LENGTH = 255
+MAX_VARCHAR_LENGTH = 65535
+SYSTEM_DEFAULT_PRECISION = 38
+SYSTEM_DEFAULT_SCALE = 18
+USER_DEFAULT_PRECISION = 10
+USER_DEFAULT_SCALE = 0
+MAX_PRECISION = 38
+MAX_SCALE = 38
