+++ storage-api/serde-api/src/gen/thrift/gen-rb/serde_constants.rb
@@ -133,3 +133,19 @@ IntegralTypes = Set.new([
   %q"bigint",
 ])
 
+MAX_CHAR_LENGTH = 255
+
+MAX_VARCHAR_LENGTH = 65535
+
+SYSTEM_DEFAULT_PRECISION = 38
+
+SYSTEM_DEFAULT_SCALE = 18
+
+USER_DEFAULT_PRECISION = 10
+
+USER_DEFAULT_SCALE = 0
+
+MAX_PRECISION = 38
+
+MAX_SCALE = 38
+
